moved glue out of main
